+++ widgets/src/lib/components/DemoThemeSwitcher/DemoThemeSwitcher.svelte
